use StorageManager directly instead of Workspace to remove need for ES

+++ cms/views.py
@@ -3,7 +3,7 @@ import os.path
 import shutil
 
 from django.core.exceptions import ValidationError
-from django.http import HttpResponse
+from django.http import HttpResponse, HttpResponseNotAllowed
 from django.shortcuts import redirect
 from django.conf import settings
 from django.contrib.auth.decorators import login_required
@@ -11,6 +11,7 @@ from django.views.decorators.csrf import csrf_exempt
 
 from git import Repo
 from elasticgit import EG
+from elasticgit.storage import StorageManager
 
 from cms import models, utils
 from cms.management.commands.import_from_git import Command
@@ -19,10 +20,15 @@ from unicore.content.models import (
     Category, Page, Localisation as EGLocalisation)
 
 
-def clone_repo(url, name):
+def clone_repo(url, name, delete_if_exists=True):
     repo_path = os.path.join(settings.IMPORT_CLONE_REPO_PATH, name)
-    if os.path.exists(repo_path):
-        shutil.rmtree(repo_path)
+    if delete_if_exists:
+        if os.path.exists(repo_path):
+            shutil.rmtree(repo_path)
+    elif os.path.exists(repo_path):
+        repo = Repo(repo_path)
+        StorageManager(repo).pull()
+        return repo
     return Repo.clone_from(url, repo_path)
 
 
@@ -36,46 +42,44 @@ def import_clone_repo(request, *args, **kwargs):
                 'Invalid repo_url',
                 status=400,
                 mimetype='application/json')
-        repo_index = 'import-repo-prefix-%s' % utils.parse_repo_name(url)
-        repo = clone_repo(url, repo_index)
-        ws = utils.setup_workspace(repo.working_dir, repo_index)
-        ws.sync(EGLocalisation)
-        ws.sync(Category)
-        ws.sync(Page)
-
+        repo_name = 'import-repo-prefix-%s' % utils.parse_repo_name(url)
+        repo = clone_repo(url, repo_name)
+        # NOTE: we can iterate over all instances of a model
+        # by using the elastic-git StorageManager directly
         localisations = [
-            l.to_object().locale for l in ws.S(EGLocalisation).everything()]
-
+            l.locale for l in
+            StorageManager(repo).iterate(EGLocalisation)]
         return HttpResponse(
-            json.dumps({'locales': localisations, 'index_prefix': repo_index}),
+            json.dumps({'locales': localisations, 'repo_name': repo_name}),
             mimetype='application/json')
     return redirect('/github/import/choose/')
 
 
-def import_locale_content(workspace, locale):
+def import_locale_content(repo, locales):
     mngmnt_command = Command()
+    sm = StorageManager(repo)
 
-    [l] = workspace.S(EGLocalisation).filter(locale=locale)
-    l = l.to_object()
-    language_code, _, country_code = l.locale.partition('_')
-    localisation, new = models.Localisation.objects.get_or_create(
-        language_code=language_code,
-        country_code=country_code,
-        defaults={
-            'logo_text': l.logo_text,
-            'logo_description': l.logo_description,
-        })
+    # NOTE: we can iterate over all instances of a model
+    # by using the elastic-git StorageManager directly
+    for l in sm.iterate(EGLocalisation):
+        if l.locale not in locales:
+            continue
+        language_code, _, country_code = l.locale.partition('_')
+        localisation, new = models.Localisation.objects.get_or_create(
+            language_code=language_code,
+            country_code=country_code,
+            defaults={
+                'logo_text': l.logo_text,
+                'logo_description': l.logo_description,
+            })
 
-    if new:
-        mngmnt_command.set_image_field(l, localisation, 'image')
-        mngmnt_command.set_image_field(l, localisation, 'logo_image')
+        if new:
+            mngmnt_command.set_image_field(l, localisation, 'image')
+            mngmnt_command.set_image_field(l, localisation, 'logo_image')
 
-    workspace.refresh_index()
-    categories = workspace.S(Category).filter(language=locale)[:1000]
-
-    for instance in categories:
-        instance = instance.to_object()
-
+    for instance in sm.iterate(Category):
+        if instance.language not in locales:
+            continue
         category, _ = models.Category.objects.get_or_create(
             uuid=instance.uuid,
             defaults={
@@ -91,14 +95,9 @@ def import_locale_content(workspace, locale):
         mngmnt_command.set_image_field(instance, category, 'image')
         category.save()
 
-    # Manually refresh stuff because the command disables signals
-    workspace.refresh_index()
-
-    pages = workspace.S(Page).filter(language=locale)[:1000]
-
-    for instance in pages:
-        instance = instance.to_object()
-
+    for instance in sm.iterate(Pages):
+        if instance.language not in locales:
+            continue
         primary_category = None
         if instance.primary_category:
             primary_category = models.Category.objects.get(
@@ -123,7 +122,6 @@ def import_locale_content(workspace, locale):
                     'position': instance.position or 0
                 }
             )
-            workspace.refresh_index()
             # add the tags
             post.author_tags.add(*instance.author_tags)
 
@@ -133,13 +131,10 @@ def import_locale_content(workspace, locale):
             print 'An error occured with: %s(%s)' % (
                 instance.title, instance.uuid)
 
-    # Manually refresh stuff because the command disables signals
-    workspace.refresh_index()
-
     # second pass to add related fields
-    for instance in pages:
-        instance = instance.to_object()
-
+    for instance in sm.iterate(Pages):
+        if instance.language not in locales:
+            continue
         if instance.linked_pages:
             p = models.Post.objects.get(uuid=instance.uuid)
             p.related_posts.add(*list(
@@ -150,20 +145,25 @@ def import_locale_content(workspace, locale):
 @login_required
 def import_repo(request, *args, **kwargs):
     if request.is_ajax():
-        index_prefix = request.POST.get('index_prefix')
+        repo_name = request.POST.get('repo_name')
         locales = request.POST.getlist('locales[]')
 
-        repo_path = os.path.join(settings.IMPORT_CLONE_REPO_PATH, index_prefix)
-        workspace = EG.workspace(
-            repo_path, index_prefix=index_prefix,
-            es={'urls': [settings.ELASTICSEARCH_HOST]})
-
-        for locale in locales:
-            import_locale_content(workspace, locale)
-
-        workspace.destroy()
+        repo_path = os.path.join(settings.IMPORT_CLONE_REPO_PATH, repo_name)
+        repo = Repo(repo_path)
+        import_locale_content(repo, locales)
+        shutil.rmtree(repo_path)
 
         return HttpResponse(
             json.dumps({'success': True}),
             mimetype='application/json')
     return redirect('/github/import/choose/')
+
+
+@csrf_exempt
+def import_repo_hook(request):
+    if request.method != 'POST':
+        return HttpResponseNotAllowed(['POST'])
+
+    # TODO: authorize request by checking domain or CAS token
+
+    return HttpResponse()
